refactor : 일관성에 맞게 디렉터리 구조 변경
