+++ module-batch/src/main/java/com/example/modulebatch/count/job/MakeFeedStatisticsJobConfig.java
@@ -18,7 +18,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import com.example.e2ekernelengine.domain.feed.db.entity.Feed;
-import com.example.modulebatch.count.repository.FeedStatistics;
+import com.example.modulebatch.count.db.entity.FeedStatistics;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
